+++ lib/widgets/desktop_shell.dart
@@ -137,6 +137,12 @@ class _DesktopRail extends StatefulWidget {
       Icons.layers,
       'Treble',
     ),
+    _RailDest(
+      AppRoutes.community,
+      Icons.forum_outlined,
+      Icons.forum,
+      'Community',
+    ),
   ];
 
   @override